requirements.txt
# Database
psycopg2-binary==2.9.9

# Configuration
python-dotenv==1.0.1

# CLI
click==8.1.7

# Testing
pytest==8.0.0
pytest-mock==3.12.0

# Utilities
python-dateutil==2.8.2
